requirements.txt
zope.configuration>=4.1.0
zope.interface>=4.5.0
zope.component>=4.4.1
maya>=0.5.0
z3c.autoinclude>=0.3.7
cachetools>=2.1.0
python-gnupg>=0.4.2
attrs>=18.1.0

